fix: image renaming
