+++ packages/gatsby-theme-docs/src/components/markdown-fragment-image.js
