+++ packages/gatsby-theme-docs/src/utils/markdown-fragment-to-react.js
@@ -6,7 +6,7 @@ import remark2react from 'remark-react';
 import frontmatter from 'remark-frontmatter';
 import { designSystem, Markdown } from '@commercetools-docs/ui-kit';
 import Link from '../components/link';
-import MarkdownFragmentImage from '../components/image';
+import MarkdownFragmentImage from '../components/markdown-fragment-image';
 
 const Div = styled.div``;
 const Heading = styled.p`
